Prototype v1 completed
Enemies now chase flag and return it to the FlagGoal. Game Over is tied to the Music and the MusicPlayer. Pause Menu is setup to accept multiplayer and will ignore Esc if Multiplayer is set to True. There is still a slight delay between the Pause button pressed and the actual music stopping.